refactored clipboard service
