+++ src/app/navigation/components/site-layout/site-layout.component.html
@@ -45,7 +45,6 @@
         <mat-icon class="btn-icon">
           {{ link.iconName }}
         </mat-icon>
-        <span *ngIf="showFiller">{{ link.name | titlecase }}</span>
       </button>
     </ng-container>
   </div>
