+++ src/app/navigation/components/site-layout/site-layout.component.ts
@@ -2,8 +2,8 @@ import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { map } from 'rxjs';
-import { headerUserNavigationMenuLinks } from 'src/app/navigation/models/header-user-navigation-menu-links';
 
+import { headerUserNavigationMenuLinks } from 'src/app/navigation/models/header-user-navigation-menu-links';
 import { userActions } from 'src/app/user/state/actions';
 import { selectSiteNavigationLinksDataByUserRole } from 'src/app/user/state/selectors';
 
@@ -20,7 +20,9 @@ export class SiteLayoutComponent implements OnInit {
 
   headerUserNavigationMenuLinksData$ = this.siteNavigationLinksData$.pipe(
     map((linksData) =>
-      linksData?.filter((link) => headerUserNavigationMenuLinks.includes(link.name)),
+      linksData?.filter((link) =>
+        headerUserNavigationMenuLinks.includes(link.name),
+      ),
     ),
   );
