+++ src/app/navigation/models/site-navigation-link.ts
@@ -1,5 +1,5 @@
 import { SiteNavigationLinkNames } from 'src/app/navigation/models/site-navigation-links-names';
-import { UserRoles } from 'src/app/shared/constants/constants';
+import { UserRoles } from 'src/app/shared/constants/user-roles';
 
 export interface ISiteNavigationLink {
   path: string[];
