--- src/app/shared/services/clipboard/clipboard-property-names-registry.ts
@@ -1,45 +1,0 @@
-export const clipboardPropertyNamesRegistry = new Map<string, string>([
-  // user personal info
-  ['email', 'Email'],
-  ['name', 'Name'],
-  ['surname', 'Surname'],
-  ['positionDescription', 'Position description'],
-  ['language', 'Language'],
-  ['endReason', 'Dismissal reason'],
-  ['phone', 'Phone number'],
-  ['salary', 'Salary'],
-  ['skills', 'Skills'],
-  ['experience', 'Experience'],
-  ['isBanned', ' Is Banned'],
-  ['banReason', 'Ban Reason'],
-  ['birthday', 'Birthday'],
-  ['startDate', 'Start Day'],
-  ['endDate', 'Dismissal Day'],
-  ['status', 'Status'],
-
-  // contact details
-  ['city', 'City'],
-  ['address', 'Address'],
-  ['', ''],
-
-  //social networks
-  ['upwork', 'Upwork Link'],
-  ['github', 'Github Link'],
-  ['linkedin', 'Linkedin Link'],
-  ['telegramTag', 'Telegram Tag'],
-
-  // education information
-  ['universityName', 'University Name'],
-  ['specialization', 'Specialization'],
-  ['pricingModel', 'Degree'],
-
-  //user bank info
-  ['individualEntrepreneurName', 'Individual Entrepreneur Name'],
-  ['individualEntrepreneurAddress', 'Individual Entrepreneur Address'],
-  ['individualEntrepreneurIndividualTaxNumber', 'Individual Tax Number'],
-  ['individualEntrepreneurBankAccounNumber', 'Individual Bank Account Number'],
-  ['individualEntrepreneurBankName', 'Bank Name'],
-  ['individualEntrepreneurBankCode', 'Bank Code'],
-  ['individualEntrepreneurBeneficiaryBank', 'Beneficiary Bank Name'],
-  ['individualEntrepreneurSwiftCode', 'SWIFT Code'],
-]);
