+++ src/app/user/components/user-profile/constants/clipboard-property-names-registries/clipboard-bank-socials-registy.ts
@@ -1,0 +1,14 @@
+export const clipboardBankSocialsRegistry = new Map<string, string>([
+  ['individualEntrepreneurName', 'Individual Entrepreneur Name'],
+  ['individualEntrepreneurAddress', 'Individual Entrepreneur Address'],
+  ['individualEntrepreneurIndividualTaxNumber', 'Individual Tax Number'],
+  ['individualEntrepreneurBankAccounNumber', 'Individual Bank Account Number'],
+  ['individualEntrepreneurBankName', 'Bank Name'],
+  ['individualEntrepreneurBankCode', 'Bank Code'],
+  ['individualEntrepreneurBeneficiaryBank', 'Beneficiary Bank Name'],
+  ['individualEntrepreneurSwiftCode', 'SWIFT Code'],
+  ['upwork', 'Upwork Link'],
+  ['github', 'Github Link'],
+  ['linkedin', 'Linkedin Link'],
+  ['telegramTag', 'Telegram Tag'],
+]);
