+++ src/app/user/components/user-profile/constants/clipboard-property-names-registries/clipboard-education-contacts-registry.ts
@@ -1,0 +1,10 @@
+export const clipboardEducationContactsRegistry = new Map<string, string>([
+  ['city', 'City'],
+  ['address', 'Address'],
+  ['universityName', 'University Name'],
+  ['specialization', 'Specialization'],
+  ['pricingModel', 'Degree'],
+  ['startDate', 'Entrance Date'],
+  ['endDate', 'Graduation Date'],
+  ['status', 'Status'],
+]);
