+++ src/app/user/components/user-profile/constants/clipboard-property-names-registries/clipboard-personal-info-registry.ts
@@ -1,0 +1,18 @@
+export const clipboardPersonalInfoRegistry = new Map<string, string>([
+  ['email', 'Email'],
+  ['name', 'Name'],
+  ['surname', 'Surname'],
+  ['positionDescription', 'Position description'],
+  ['language', 'Language'],
+  ['endReason', 'Dismissal reason'],
+  ['phone', 'Phone number'],
+  ['salary', 'Salary'],
+  ['skills', 'Skills'],
+  ['experience', 'Experience'],
+  ['isBanned', ' Is Banned'],
+  ['banReason', 'Ban Reason'],
+  ['birthday', 'Birthday'],
+  ['startDate', 'Start Day'],
+  ['endDate', 'Dismissal Day'],
+  ['status', 'Status'],
+]);
