+++ src/app/user/components/user-profile/constants/edit-dialog-options.ts
@@ -1,6 +1,4 @@
 export const editDialogOptions = {
   restoreFocus: false,
   autoFocus: false,
-  // width: '100vw',
-  // maxWidth: '100vw',
 };
