+++ src/app/user/components/user-profile/user-edit/user-edit.scss
@@ -20,7 +20,6 @@
 
 .inner {
   display: grid;
-
   column-gap: $m-size;
   row-gap: $xs-size;
