--- src/app/user/components/user-profile/user-education-contacts/user-education-contacts.component.scss
@@ -1,1 +1,0 @@
-
