+++ src/app/user/components/user-profile/user-education-contacts/user-education-contacts.component.ts
@@ -5,6 +5,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { UserStore } from 'src/app/user/components/user-profile/user-profile.store';
 import { EditContactsModalComponent } from 'src/app/user/components/user-profile/user-edit/edit-contacts/edit-contacts-modal.component';
 import { ClipboardService } from 'src/app/shared/services/clipboard/clipboard.service';
+import { clipboardEducationContactsRegistry } from 'src/app/user/components/user-profile/constants/clipboard-property-names-registries/clipboard-education-contacts-registry';
 
 @Component({
   selector: 'app-user-education-contacts',
@@ -24,6 +25,8 @@ export class UserEducationContactsComponent {
       city: user?.city,
     })),
   );
+
+  clipboardRegistry = clipboardEducationContactsRegistry;
 
   constructor(
     private readonly _userStore: UserStore,
