+++ src/app/user/components/user-profile/user-personal-info/user-personal-info.component.html
@@ -18,7 +18,7 @@
         mat-icon-button
         class="icon-btn"
         aria-label="Copy content button"
-        (click)="clipboardService.copyToClipboard(userData)">
+        (click)="clipboardService.copyToClipboard(userData, clipboardRegistry)">
         <mat-icon>content_copy</mat-icon>
       </button>
     </div>
