+++ src/app/user/components/user-profile/user-personal-info/user-personal-info.scss
@@ -6,19 +6,18 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-
   gap: $m-size;
-  padding: $m-size 0;
   border-bottom: $base-border;
 
   @include media(tablet) {
-    padding: 0 $m-size;
+    padding: $m-size;
     flex-direction: row;
     border: none;
   }
 }
 .inner {
-  width: 90%;
+  width: 100%;
+
   @include media(tablet) {
     width: auto;
     flex-grow: 1;
@@ -38,7 +37,13 @@
   display: flex;
   flex-grow: 1;
   justify-content: space-between;
+  padding: $m-size;
+
+  @include media(tablet) {
+    padding: 0;
+  }
 }
+
 .terms {
   align-self: flex-end;
 }
@@ -55,6 +60,7 @@
   display: flex;
   align-items: center;
   gap: $xs-size;
+  margin-bottom: $xs-size;
 
   @include media(tablet) {
     justify-content: flex-start;
@@ -69,8 +75,9 @@
 .detail {
   display: flex;
   align-items: center;
-  gap: $xs-size;
+  gap: $xxs-size;
   margin-bottom: $xxs-size;
+  min-height: 24px;
 }
 
 .term {
