+++ src/app/user/components/user-profile/user-profile.component.ts
@@ -9,6 +9,7 @@ import {
 } from 'src/app/shared/constants/media-width';
 
 import { ClipboardService } from 'src/app/shared/services/clipboard/clipboard.service';
+import { clipboardPersonalInfoRegistry } from 'src/app/user/components/user-profile/constants/clipboard-property-names-registries/clipboard-personal-info-registry';
 import { editDialogOptions } from 'src/app/user/components/user-profile/constants/edit-dialog-options';
 import { EditDescriptionModalComponent } from 'src/app/user/components/user-profile/user-edit/edit-description-modal/edit-description-modal.component';
 import { UserStore } from 'src/app/user/components/user-profile/user-profile.store';
@@ -23,6 +24,8 @@ import { UserStore } from 'src/app/user/components/user-profile/user-profile.sto
 export class UserProfileComponent implements OnInit {
   user$ = this._userStore.user$;
   loading$ = this._userStore.loading$;
+
+  clipboardRegistry = clipboardPersonalInfoRegistry;
 
   constructor(
     private route: ActivatedRoute,
