+++ src/styles/_media-mixins.scss
@@ -25,12 +25,6 @@
     }
   }
 
-  @if $type == 'not-desktop' {
-    @media screen and (max-width: #{$desktop - 1}) {
-      @content;
-    }
-  }
-
   @if $type == 'retina' {
     @media (min-device-pixel-ratio: 2),
       (min-resolution: 192dpi),
