+++ tsconfig.json
@@ -9,7 +9,6 @@
     "strict": true,
     "strictNullChecks": true,
     "noImplicitOverride": true,
-    // "noImplicitAny": false,
     "noPropertyAccessFromIndexSignature": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
